requirements.txt
cryptg==0.5.0.post0
mysql-connector-python==9.2.0
pillow==11.1.0
pyaes==1.6.1
pyasn1==0.6.1
python-dotenv==1.1.0
rsa==4.9
setuptools==78.1.0
Telethon==1.39.0
